+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - hel-965: Ajout du sous texte dans la catégorie Finance de la selection d’indicateurs sur la page de comparaison
 - hel-956: Montée de version NextJS en v16
+- hel-964: Ajout de l’indicateur TO occupation global dans la page de comparaison
 
 ## [1.7.8](https://github.com/DNUM-SocialGouv/helios/compare/release/1.7.7...release/1.7.8) (2026-01-19)